Rename tests -> abel_tests
