+++ tools/abel/abel_tests/entity_field_test.py
@@ -18,12 +18,12 @@ from absl.testing import absltest
 from model.entity_field import EntityField
 from model.state import State
 from model.units import Units
-from abel.tests.test_constants import TEST_ENTITY_FIELD_DICT_WITH_STATES
-from abel.tests.test_constants import TEST_ENTITY_FIELD_DICT_WITH_UNITS
-from abel.tests.test_constants import TEST_RAW_FIELD_NAME
-from abel.tests.test_constants import TEST_STANDARD_FIELD_NAME
-from abel.tests.test_constants import TEST_STATE_DICT
-from abel.tests.test_constants import TEST_UNITS
+from abel_tests.test_constants import TEST_ENTITY_FIELD_DICT_WITH_STATES
+from abel_tests.test_constants import TEST_ENTITY_FIELD_DICT_WITH_UNITS
+from abel_tests.test_constants import TEST_RAW_FIELD_NAME
+from abel_tests.test_constants import TEST_STANDARD_FIELD_NAME
+from abel_tests.test_constants import TEST_STATE_DICT
+from abel_tests.test_constants import TEST_UNITS
 
 
 class EntityFieldTest(absltest.TestCase):
