+++ tools/abel/abel_tests/export_helper_test.py
@@ -28,13 +28,14 @@ from model.export_helper import GoogleSheetExport
 from model.guid_to_entity_map import GuidToEntityMap
 from model.model_builder import ModelBuilder
 from model.model_error import SpreadsheetAuthorizationError
-from abel.tests.test_constants import TEST_RESOURCES
-from abel.tests.test_constants import TEST_SPREADSHEET
+from abel_tests.test_constants import TEST_RESOURCES
+from abel_tests.test_constants import TEST_SPREADSHEET
 
 _GOOD_TEST_BUILDING_CONFIG = os.path.join(TEST_RESOURCES,
                                           'good_test_building_config.yaml')
 
 
+# pylint: disable=consider-using-with
 class ExportHelperTest(absltest.TestCase):
   """Tests export methods for ABEL."""
