+++ tools/abel/abel_tests/guid_to_entity_map_test.py
@@ -19,11 +19,11 @@ from model.entity import ReportingEntity
 from model.entity import VirtualEntity
 from model.guid_to_entity_map import GuidToEntityMap
 from model.site import Site
-from abel.tests.test_constants import TEST_REPORTING_ENTITY_DICT
-from abel.tests.test_constants import TEST_REPORTING_GUID
-from abel.tests.test_constants import TEST_SITE_DICT
-from abel.tests.test_constants import TEST_SITE_DICT_NO_GUID
-from abel.tests.test_constants import TEST_VIRTUAL_ENTITY_DICT
+from abel_tests.test_constants import TEST_REPORTING_ENTITY_DICT
+from abel_tests.test_constants import TEST_REPORTING_GUID
+from abel_tests.test_constants import TEST_SITE_DICT
+from abel_tests.test_constants import TEST_SITE_DICT_NO_GUID
+from abel_tests.test_constants import TEST_VIRTUAL_ENTITY_DICT
 
 _TEST_REPORTING_ENTITY = ReportingEntity.FromDict(TEST_REPORTING_ENTITY_DICT)
 _TEST_VIRTUAL_ENTITY = VirtualEntity.FromDict(TEST_VIRTUAL_ENTITY_DICT)
