+++ tools/abel/abel_tests/model_builder_test.py
@@ -29,14 +29,14 @@ from model.guid_to_entity_map import GuidToEntityMap
 from model.model_builder import ModelBuilder
 from model.site import Site
 from model.state import State
-from abel.tests.test_constants import TEST_CONNECTION_DICT
-from abel.tests.test_constants import TEST_ENTITY_FIELD_DICT_WITH_STATES
-from abel.tests.test_constants import TEST_ENTITY_FIELD_DICT_WITH_UNITS
-from abel.tests.test_constants import TEST_REPORTING_ENTITY_DICT
-from abel.tests.test_constants import TEST_RESOURCES
-from abel.tests.test_constants import TEST_SITE_DICT
-from abel.tests.test_constants import TEST_STATE_DICT
-from abel.tests.test_constants import TEST_VIRTUAL_ENTITY_DICT
+from abel_tests.test_constants import TEST_CONNECTION_DICT
+from abel_tests.test_constants import TEST_ENTITY_FIELD_DICT_WITH_STATES
+from abel_tests.test_constants import TEST_ENTITY_FIELD_DICT_WITH_UNITS
+from abel_tests.test_constants import TEST_REPORTING_ENTITY_DICT
+from abel_tests.test_constants import TEST_RESOURCES
+from abel_tests.test_constants import TEST_SITE_DICT
+from abel_tests.test_constants import TEST_STATE_DICT
+from abel_tests.test_constants import TEST_VIRTUAL_ENTITY_DICT
 
 
 class ModelBuilderTest(absltest.TestCase):
