+++ tools/abel/abel_tests/test_resources/good_test_building_config.yaml
